Confusion color scheme added
Now there is the option to visualize the instances according to whether
or not they are TP (green), TN (grey), FP (yellow), or FN (red). One
can do this for each class by clicking the rectangle of the respective
class on the legend located at the upper left corner of the screen. To
go back to the regular visualization mode where the colors represent
the correct class one should double click any of the aforementioned
rectangles.